[update] loan amount to int, nullable until admin approval, add member_chat_id

+++ account/service.py
@@ -10,13 +10,13 @@ class AccountService:
     def __init__(self, db: SessionLocal = None):
         self.db: SessionLocal = db or get_db()
 
-    def credit(self, user_id: int, amount: float):
+    def credit(self, user_id: int, amount: int):
         account = self.db.query(Account).filter_by(user_id=user_id).first()
         if not account:
             account = Account(user_id=user_id, balance=0)
             self.db.add(account)
             self.db.flush()
-        account.balance = float(account.balance) + float(amount)
+        account.balance = int(account.balance) + amount
 
     def get_balance(self, data: dict):
         account = self.db.query(Account).filter_by(user_id=data["user_id"]).first()
